backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-Levenshtein==0.25.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
onnxruntime==1.17.0
aiofiles==23.2.1
httpx==0.26.0
